refactor: moved adbc traits to sync mod
